Test: refactoring item.test.js separating functions in seed.js file.
