+++ tests/unit/server/seed.js
@@ -1,0 +1,20 @@
+const Item = require("../../../server/models/item");
+
+const seedItems = [
+  {
+    title: "Test item 1"
+  },
+  {
+    title: "Test item 2"
+  }
+];
+
+const populateItems = async () => {
+  await Item.deleteMany();
+  await Item.insertMany(seedItems);
+};
+
+module.exports = {
+  seedItems,
+  populateItems
+};
